pyproject.toml
[project]
name = "crackhash-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "markdown>=3.10.2",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
]
